+++ Compiler/src/edu/mit/compilers/st/Manager.java
@@ -182,7 +182,7 @@ public class Manager {
     public static final void tmpPush(Oprand tmp) {
         if (tmp instanceof Reg) {
             callerSavedRegsUsage.put(((Reg)tmp).getRegName(), ((Reg)tmp));
-        } else if (tmp instanceof Addr) {
+        } else if (tmp instanceof RegAddr) {
             ((Addr)tmp).getReservedRegs().forEach(e -> callerSavedRegsUsage.put(e.getRegName(), e));
         }
         tmpStack.push(tmp);
@@ -192,7 +192,7 @@ public class Manager {
         Oprand returnOp = tmpStack.pop();
         if (returnOp instanceof Reg) {
             callerSavedRegsUsage.remove(((Reg)returnOp).getRegName());
-        } else if (returnOp instanceof Addr) {
+        } else if (returnOp instanceof RegAddr) {
             ((Addr)returnOp).getReservedRegs().forEach(e -> callerSavedRegsUsage.remove(e.getRegName()));
         }
         return returnOp;
